+++ swarm/src/connection.rs
@@ -26,8 +26,6 @@ pub use error::{
     ConnectionError, PendingConnectionError, PendingInboundConnectionError,
     PendingOutboundConnectionError,
 };
-pub use pool::{ConnectionCounters, ConnectionLimits};
-pub use pool::{EstablishedConnection, PendingConnection};
 
 use crate::handler::ConnectionHandler;
 use crate::upgrade::{InboundUpgradeSend, OutboundUpgradeSend, SendWrapper};